Prepare company app for v4 version of company API
