+++ config/api_urls.py
@@ -33,7 +33,7 @@ v3_urls = [
         ),
     ),
     path('', include((company_urls.contact_urls, 'contact'), namespace='contact')),
-    path('', include((company_urls.company_urls, 'company'), namespace='company')),
+    path('', include((company_urls.company_urls_v3, 'company'), namespace='company')),
     path('', include((company_urls.ch_company_urls, 'ch-company'), namespace='ch-company')),
     path('', include((event_urls, 'event'), namespace='event')),
     path('', include((feature_flag_urls, 'feature-flag'), namespace='feature-flag')),
@@ -49,3 +49,8 @@ v3_urls = [
         ),
     ),
 ]
+
+
+# API V4 - new format for addresses
+
+v4_urls = []
