+++ datahub/company/test/test_ch_company_views.py
@@ -1,0 +1,77 @@
+from rest_framework import status
+from rest_framework.reverse import reverse
+
+from datahub.company.models import CompaniesHouseCompany
+from datahub.company.test.factories import CompaniesHouseCompanyFactory
+from datahub.core.test_utils import APITestMixin, format_date_or_datetime
+
+
+class TestCHCompany(APITestMixin):
+    """CH company tests."""
+
+    def test_list_ch_companies(self):
+        """List the companies house companies."""
+        CompaniesHouseCompanyFactory()
+        CompaniesHouseCompanyFactory()
+
+        url = reverse('api-v3:ch-company:collection')
+        response = self.api_client.get(url)
+
+        assert response.status_code == status.HTTP_200_OK
+        assert response.json()['count'] == CompaniesHouseCompany.objects.all().count()
+
+    def test_get_ch_company(self):
+        """Test retrieving a single CH company."""
+        ch_company = CompaniesHouseCompanyFactory()
+        url = reverse(
+            'api-v3:ch-company:item', kwargs={'company_number': ch_company.company_number},
+        )
+        response = self.api_client.get(url)
+
+        assert response.status_code == status.HTTP_200_OK
+        response_data = response.json()
+        assert response_data == {
+            'id': ch_company.id,
+            'business_type': {
+                'id': str(ch_company.business_type.id),
+                'name': ch_company.business_type.name,
+            },
+            'company_number': ch_company.company_number,
+            'company_category': ch_company.company_category,
+            'company_status': ch_company.company_status,
+            'incorporation_date': format_date_or_datetime(ch_company.incorporation_date),
+            'name': ch_company.name,
+            'registered_address_1': ch_company.registered_address_1,
+            'registered_address_2': ch_company.registered_address_2,
+            'registered_address_town': ch_company.registered_address_town,
+            'registered_address_county': ch_company.registered_address_county,
+            'registered_address_postcode': ch_company.registered_address_postcode,
+            'registered_address_country': {
+                'id': str(ch_company.registered_address_country.id),
+                'name': ch_company.registered_address_country.name,
+            },
+            'sic_code_1': ch_company.sic_code_1,
+            'sic_code_2': ch_company.sic_code_2,
+            'sic_code_3': ch_company.sic_code_3,
+            'sic_code_4': ch_company.sic_code_4,
+            'uri': ch_company.uri,
+        }
+
+    def test_get_ch_company_alphanumeric(self):
+        """Test retrieving a single CH company where the company number contains letters."""
+        CompaniesHouseCompanyFactory(company_number='SC00001234')
+        url = reverse(
+            'api-v3:ch-company:item', kwargs={'company_number': 'SC00001234'},
+        )
+        response = self.api_client.get(url)
+
+        assert response.status_code == status.HTTP_200_OK
+        response_data = response.json()
+        assert response_data['company_number'] == 'SC00001234'
+
+    def test_ch_company_cannot_be_written(self):
+        """Test CH company POST is not allowed."""
+        url = reverse('api-v3:ch-company:collection')
+        response = self.api_client.post(url)
+
+        assert response.status_code == status.HTTP_403_FORBIDDEN
