+++ datahub/dbmaintenance/management/commands/update_company_global_hq.py
@@ -1,7 +1,7 @@
 import reversion
 
 from datahub.company.models import Company
-from datahub.company.serializers import CompanySerializer
+from datahub.company.serializers import CompanySerializerV3
 from datahub.dbmaintenance.management.base import CSVBaseCommand
 from datahub.dbmaintenance.utils import parse_uuid
 
@@ -42,7 +42,7 @@ class Command(CSVBaseCommand):
                 'global_headquarters': global_hq,
             }
 
-            serializer = CompanySerializer(
+            serializer = CompanySerializerV3(
                 instance=company,
                 data=data,
                 partial=True,
